zmena id z long na int aj na serveri, uprava repetition repository aby sa vracalo null ak sa nenajde repetition

+++ ServerProject/src/main/java/com/example/vavaplanit/database/mappers/EventMapper.java
@@ -22,7 +22,7 @@ public class EventMapper {
             LocalTime ends = resultSet.getObject("ends", LocalTime.class);
             LocalDate alertDate = resultSet.getObject("alert_date", LocalDate.class);
             LocalTime alert = resultSet.getObject("alert", LocalTime.class);
-            Long exceptionId = resultSet.getLong("exception_id");
+            Integer exceptionId = resultSet.getInt("exception_id");
             return new Event(
                     idEvent,
                     exceptionId,
