+++ ServerProject/src/main/java/com/example/vavaplanit/model/repetition/MonthlyRepetition.java
@@ -23,7 +23,7 @@ public class MonthlyRepetition extends WeeklyRepetition {
     public MonthlyRepetition() {
     }
 
-    public MonthlyRepetition(Long eventId, LocalDate start, LocalDate end, int repetitionInterval,
+    public MonthlyRepetition(Integer eventId, LocalDate start, LocalDate end, int repetitionInterval,
                              Integer daysOfWeek, Integer dayOfMonth, Integer ordinal) {
         super(eventId, start, end, repetitionInterval, daysOfWeek);
         this.dayOfMonth = dayOfMonth;
