+++ ServerProject/src/test/java/com/example/vavaplanit/model/repetition/MonthlyRepetitionTest.java
@@ -25,7 +25,7 @@ public class MonthlyRepetitionTest {
             public void monthlyRepetitionEveryMonthOn20th() {
                 LocalDate expectedDate = LocalDate.of(2020, 8, 20);
 
-                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1L, LocalDate.of(2020, 6, 20),
+                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1, LocalDate.of(2020, 6, 20),
                         LocalDate.of(2020, 12, 20), 1, null, 20,
                         null, 8, 2020).get(0);
 
@@ -36,7 +36,7 @@ public class MonthlyRepetitionTest {
             public void monthlyRepetitionEveryMonthOn31st() {
                 LocalDate expectedDate = LocalDate.of(2020, 2, 29);
 
-                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1L, LocalDate.of(2020, 1, 31),
+                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1, LocalDate.of(2020, 1, 31),
                         LocalDate.of(2020, 12, 31), 1, null, 31,
                         null, 2, 2020).get(0);
 
@@ -47,7 +47,7 @@ public class MonthlyRepetitionTest {
             public void monthlyRepetitionEverySecondMonthWithoutOccurrenceInMonth() {
                 List<LocalDate> expectedDate = new ArrayList<>();
 
-                List<LocalDate> actualDate = createMonthlyRepetitionAndGetDates(1L, LocalDate.of(2020, 1, 31),
+                List<LocalDate> actualDate = createMonthlyRepetitionAndGetDates(1, LocalDate.of(2020, 1, 31),
                         LocalDate.of(2020, 12, 31), 2, null, 31,
                         null, 2, 2020);
 
@@ -58,7 +58,7 @@ public class MonthlyRepetitionTest {
             public void monthlyRepetitionEvery13thMonthWithOccurrenceInMonth() {
                 LocalDate expectedDate = LocalDate.of(2020, 8, 15);
 
-                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1L, LocalDate.of(2019, 7, 15),
+                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1, LocalDate.of(2019, 7, 15),
                         LocalDate.of(2020, 12, 15), 13, null, 15,
                         null, 8, 2020).get(0);
 
@@ -73,7 +73,7 @@ public class MonthlyRepetitionTest {
             public void monthlyRepetitionEveryMonthOnLastMonday() {
                 LocalDate expectedDate = LocalDate.of(2020, 8, 31);
 
-                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1L, LocalDate.of(2020, 7, 27),
+                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1, LocalDate.of(2020, 7, 27),
                         LocalDate.of(2020, 12, 28), 1, 1, null,
                         5, 8, 2020).get(0);
 
@@ -84,7 +84,7 @@ public class MonthlyRepetitionTest {
             public void monthlyRepetitionEveryMonthOnFirstFriday() {
                 LocalDate expectedDate = LocalDate.of(2020, 8, 7);
 
-                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1L, LocalDate.of(2020, 7, 3),
+                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1, LocalDate.of(2020, 7, 3),
                         LocalDate.of(2020, 12, 4), 1, 16, null,
                         1, 8, 2020).get(0);
 
@@ -95,7 +95,7 @@ public class MonthlyRepetitionTest {
             public void monthlyRepetitionEveryMonthOn2ndWednesday() {
                 LocalDate expectedDate = LocalDate.of(2020, 8, 12);
 
-                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1L, LocalDate.of(2020, 7, 8),
+                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1, LocalDate.of(2020, 7, 8),
                         LocalDate.of(2020, 12, 9), 1, 4, null,
                         2, 8, 2020).get(0);
 
@@ -106,7 +106,7 @@ public class MonthlyRepetitionTest {
             public void monthlyRepetitionEvery2ndMonthOn3rdThursday() {
                 LocalDate expectedDate = LocalDate.of(2020, 8, 20);
 
-                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1L, LocalDate.of(2020, 6, 18),
+                LocalDate actualDate = createMonthlyRepetitionAndGetDates(1, LocalDate.of(2020, 6, 18),
                         LocalDate.of(2020, 12, 17), 2, 8, null,
                         3, 8, 2020).get(0);
 
@@ -117,7 +117,7 @@ public class MonthlyRepetitionTest {
             public void monthlyRepetitionEvery2ndMonthWithoutOccurrenceInMonth() {
                 List<LocalDate> expectedDate = new ArrayList<>();
 
-                List<LocalDate> actualDate = createMonthlyRepetitionAndGetDates(1L, LocalDate.of(2020, 7, 16),
+                List<LocalDate> actualDate = createMonthlyRepetitionAndGetDates(1, LocalDate.of(2020, 7, 16),
                         LocalDate.of(2020, 11, 19), 2, 8, null,
                         3, 8, 2020);
 
@@ -125,7 +125,7 @@ public class MonthlyRepetitionTest {
             }
         }
 
-        private List<LocalDate> createMonthlyRepetitionAndGetDates(Long eventId, LocalDate start, LocalDate end,
+        private List<LocalDate> createMonthlyRepetitionAndGetDates(Integer eventId, LocalDate start, LocalDate end,
                                                                    int repetitionInterval, Integer daysOfWeek, Integer dayOfMonth,
                                                                    Integer ordinal, int month, int year) {
 
@@ -144,7 +144,7 @@ public class MonthlyRepetitionTest {
         public void dateBeforeRepetitionStartReturnsFalse() {
             boolean expected = false;
 
-            boolean actual = createMonthlyRepetitionAndCheckDate(1L, LocalDate.of(2020, 9, 15),
+            boolean actual = createMonthlyRepetitionAndCheckDate(1, LocalDate.of(2020, 9, 15),
                     LocalDate.of(2020, 12, 15), 1, 2, null, 3,
                     LocalDate.of(2020, 9, 1));
 
@@ -155,7 +155,7 @@ public class MonthlyRepetitionTest {
         public void dateAfterRepetitionEndReturnsFalse() {
             boolean expected = false;
 
-            boolean actual = createMonthlyRepetitionAndCheckDate(1L, LocalDate.of(2020, 9, 15),
+            boolean actual = createMonthlyRepetitionAndCheckDate(1, LocalDate.of(2020, 9, 15),
                     LocalDate.of(2020, 12, 15), 1, null, 15, null,
                     LocalDate.of(2020, 12, 16));
 
@@ -169,7 +169,7 @@ public class MonthlyRepetitionTest {
             public void dateExactlyBetweenTwoRepetitionsReturnsFalse() {
                 boolean expected = false;
 
-                boolean actual = createMonthlyRepetitionAndCheckDate(1L, LocalDate.of(2020, 8, 15),
+                boolean actual = createMonthlyRepetitionAndCheckDate(1, LocalDate.of(2020, 8, 15),
                         LocalDate.of(2020, 12, 15), 2, null, 15, null,
                         LocalDate.of(2020, 9, 15));
 
@@ -180,7 +180,7 @@ public class MonthlyRepetitionTest {
             public void dateSomewhereBetweenTwoRepetitionsReturnsFalse() {
                 boolean expected = false;
 
-                boolean actual = createMonthlyRepetitionAndCheckDate(1L, LocalDate.of(2020, 8, 15),
+                boolean actual = createMonthlyRepetitionAndCheckDate(1, LocalDate.of(2020, 8, 15),
                         LocalDate.of(2020, 12, 15), 2, null, 15, null,
                         LocalDate.of(2020, 11, 5));
 
@@ -191,7 +191,7 @@ public class MonthlyRepetitionTest {
             public void lastDateOfMonthWithDayOfMonth31ReturnsTrue() {
                 boolean expected = true;
 
-                boolean actual = createMonthlyRepetitionAndCheckDate(1L, LocalDate.of(2020, 8, 31),
+                boolean actual = createMonthlyRepetitionAndCheckDate(1, LocalDate.of(2020, 8, 31),
                         LocalDate.of(2020, 12, 31), 1, null, 31, null,
                         LocalDate.of(2020, 9, 30));
 
@@ -202,7 +202,7 @@ public class MonthlyRepetitionTest {
             public void dateInRepetitionDayOfMonthAndInterval3ReturnsTrue() {
                 boolean expected = true;
 
-                boolean actual = createMonthlyRepetitionAndCheckDate(1L, LocalDate.of(2020, 8, 15),
+                boolean actual = createMonthlyRepetitionAndCheckDate(1, LocalDate.of(2020, 8, 15),
                         LocalDate.of(2021, 8, 15), 3, null, 15, null,
                         LocalDate.of(2020, 11, 15));
 
@@ -213,7 +213,7 @@ public class MonthlyRepetitionTest {
             public void dateInRepetitionDateWithYearOverlapReturnsTrue() {
                 boolean expected = true;
 
-                boolean actual = createMonthlyRepetitionAndCheckDate(1L, LocalDate.of(2019, 8, 31),
+                boolean actual = createMonthlyRepetitionAndCheckDate(1, LocalDate.of(2019, 8, 31),
                         LocalDate.of(2020, 12, 31), 2, null, 31, null,
                         LocalDate.of(2020, 2, 29));
 
@@ -228,7 +228,7 @@ public class MonthlyRepetitionTest {
             public void dateExactlyBetweenTwoRepetitionsReturnsFalse() {
                 boolean expected = false;
 
-                boolean actual = createMonthlyRepetitionAndCheckDate(1L, LocalDate.of(2020, 8, 15),
+                boolean actual = createMonthlyRepetitionAndCheckDate(1, LocalDate.of(2020, 8, 15),
                         LocalDate.of(2020, 12, 19), 2, 32, null, 3,
                         LocalDate.of(2020, 9, 19));
 
@@ -239,7 +239,7 @@ public class MonthlyRepetitionTest {
             public void dateSomewhereBetweenTwoRepetitionsReturnsFalse() {
                 boolean expected = false;
 
-                boolean actual = createMonthlyRepetitionAndCheckDate(1L, LocalDate.of(2020, 8, 15),
+                boolean actual = createMonthlyRepetitionAndCheckDate(1, LocalDate.of(2020, 8, 15),
                         LocalDate.of(2020, 12, 19), 2, 32, null, 3,
                         LocalDate.of(2020, 11, 5));
 
@@ -250,7 +250,7 @@ public class MonthlyRepetitionTest {
             public void lastFridayDateReturnsTrue() {
                 boolean expected = true;
 
-                boolean actual = createMonthlyRepetitionAndCheckDate(1L, LocalDate.of(2020, 8, 28),
+                boolean actual = createMonthlyRepetitionAndCheckDate(1, LocalDate.of(2020, 8, 28),
                         LocalDate.of(2020, 12, 25), 1, 16, null, 5,
                         LocalDate.of(2020, 9, 25));
 
@@ -261,7 +261,7 @@ public class MonthlyRepetitionTest {
             public void dateIn3rdWednesdayAndInterval3ReturnsTrue() {
                 boolean expected = true;
 
-                boolean actual = createMonthlyRepetitionAndCheckDate(1L, LocalDate.of(2020, 8, 19),
+                boolean actual = createMonthlyRepetitionAndCheckDate(1, LocalDate.of(2020, 8, 19),
                         LocalDate.of(2021, 5, 19), 3, 4, null, 3,
                         LocalDate.of(2020, 11, 18));
 
@@ -272,7 +272,7 @@ public class MonthlyRepetitionTest {
             public void dateInRepetitionDateWithYearOverlapReturnsTrue() {
                 boolean expected = true;
 
-                boolean actual = createMonthlyRepetitionAndCheckDate(1L, LocalDate.of(2019, 8, 31),
+                boolean actual = createMonthlyRepetitionAndCheckDate(1, LocalDate.of(2019, 8, 31),
                         LocalDate.of(2021, 2, 22), 1, 1, null, 5,
                         LocalDate.of(2021, 1, 25));
 
@@ -280,7 +280,7 @@ public class MonthlyRepetitionTest {
             }
         }
 
-        private boolean createMonthlyRepetitionAndCheckDate(Long eventId, LocalDate start, LocalDate end,
+        private boolean createMonthlyRepetitionAndCheckDate(Integer eventId, LocalDate start, LocalDate end,
                                                                    int repetitionInterval, Integer daysOfWeek, Integer dayOfMonth,
                                                                    Integer ordinal, LocalDate date) {
